+++ src/main/kotlin/io/github/lobodpav/spock/icon/SpockIcon.kt
@@ -1,0 +1,8 @@
+package io.github.lobodpav.spock.icon
+
+import com.intellij.openapi.util.IconLoader
+import javax.swing.Icon
+
+object SpockIcon {
+    val specification: Icon = IconLoader.getIcon("/icons/specification.svg", javaClass)
+}